# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Genet Open API
  description: |
    Genet OpenAPI for Pod, Deployment, StatefulSet, Service, ConfigMap, and Job operations.
    - All resources use simplified JSON request/response models.
    - Effective namespace is always derived from the API key bound `ownerUser`.
    - Server injects ownership and management labels automatically.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - BearerAuth: []
paths:
...
  /api/open/deployments:
    post:
      summary: Create Deployment
      description: |
        Create a Deployment using JSON request fields consistent with the web UI workload form.
        Resource values are per replica; total quota usage equals `replicas * per-replica resources`.
        The effective namespace is derived from API key bound `ownerUser` (user-{identifier}).
      operationId: createDeployment
      tags: [Deployments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAPIDeploymentCreateRequest"
            example:
              image: "nvidia/cuda:12.0.0-base-ubuntu22.04"
              gpuType: "NVIDIA A100"
              gpuCount: 1
              replicas: 2
              cpu: "8"
              memory: "32Gi"
              shmSize: "1Gi"
              nodeName: "gpu-node-01"
              name: "train"
      responses:
        "201":
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAPIDeploymentCreateResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: List Deployments
      operationId: listDeployments
      tags: [Deployments]
      responses:
        "200":
          description: Deployment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAPIDeploymentList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/open/deployments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Deployment ID / name

    get:
      summary: Get Deployment
      operationId: getDeployment
      tags: [Deployments]
      responses:
        "200":
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAPIDeployment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      summary: Delete Deployment
      operationId: deleteDeployment
      tags: [Deployments]
      responses:
        "200":
          description: Deployment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "API Key from admin page (preferred) or legacy config.yaml openAPI.apiKeys"
  schemas:
...
    OpenAPIUserMount:
      type: object
      properties:
        hostPath:
          type: string
          description: Host machine path to mount.
          example: "/mnt/datasets"
        mountPath:
          type: string
          description: Container mount path.
          example: "/workspace-genet/datasets"
        readOnly:
          type: boolean
          description: Whether mount is read-only.
          default: false
      required:
        - hostPath
        - mountPath
...
    OpenAPIPod:
      type: object
      properties:
        id:
          type: string
          description: Pod ID / name.
        name:
          type: string
          description: Pod name.
        namespace:
          type: string
          description: Namespace of this pod.
        container:
          type: string
          description: Main container name.
        status:
          type: string
          description: Normalized status used by UI.
        phase:
          type: string
          description: Raw Kubernetes Pod phase.
        image:
          type: string
          description: Container image.
        gpuType:
          type: string
          description: GPU type label.
        gpuCount:
          type: integer
          description: Requested GPU count.
        cpu:
          type: string
          description: CPU request.
        memory:
          type: string
          description: Memory request.
        createdAt:
          type: string
          format: date-time
          description: Pod creation timestamp.
        nodeIP:
          type: string
          description: Node IP where pod is scheduled.
        protectedUntil:
          type: string
          format: date-time
          nullable: true
          description: Protection deadline if set.
...
    OpenAPIDeploymentCreateRequest:
      type: object
      description: |
        JSON request for creating a Deployment.
        Resource values are per replica; total quota usage equals `replicas * per-replica resources`.
      properties:
        image:
          type: string
          description: Container image.
          example: "nvidia/cuda:12.0.0-base-ubuntu22.04"
        gpuType:
          type: string
          description: GPU type display name. Optional when `gpuCount` is 0.
          example: "NVIDIA A100"
        gpuCount:
          type: integer
          minimum: 0
          maximum: 8
          description: Requested GPU count per replica.
          example: 1
        replicas:
          type: integer
          minimum: 1
          maximum: 8
          description: Deployment replica count.
          example: 2
        cpu:
          type: string
          description: CPU request per replica.
          example: "8"
        memory:
          type: string
          description: Memory request per replica.
          example: "32Gi"
        shmSize:
          type: string
          description: Shared memory size per replica.
          example: "1Gi"
        nodeName:
          type: string
          description: Optional fixed scheduling node for all replicas.
          example: "gpu-node-01"
        name:
          type: string
          description: Optional custom Deployment name suffix.
          example: "train"
        userMounts:
          type: array
          description: Optional custom hostPath mounts.
          items:
            $ref: "#/components/schemas/OpenAPIUserMount"
      required:
        - image
        - gpuCount
        - replicas
    OpenAPIDeploymentCreateResponse:
      type: object
      properties:
        message:
          type: string
          description: Creation result message.
          example: "Deployment 创建成功"
        id:
          type: string
          description: Deployment ID / name.
          example: "deploy-alice-train"
        name:
          type: string
          description: Deployment name.
          example: "deploy-alice-train"
    OpenAPIDeployment:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        namespace:
          type: string
        status:
          type: string
        image:
          type: string
        gpuType:
          type: string
        gpuCount:
          type: integer
        cpu:
          type: string
        memory:
          type: string
        replicas:
          type: integer
        readyReplicas:
          type: integer
        createdAt:
          type: string
          format: date-time
        pods:
          type: array
          items:
            $ref: "#/components/schemas/OpenAPIPod"
    OpenAPIDeploymentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OpenAPIDeployment"
...
    Error:
      type: object
      properties:
        error:
          type: string
    Message:
      type: object
      properties:
        message:
          type: string
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Missing or invalid Authorization header"
    Forbidden:
      description: API key lacks required owner binding or scope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "forbidden: api key must bind ownerUser"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "resource not found"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"